requirements.txt
Package    Version
---------- -------
OPi.GPIO   0.5.2
paho-mqtt  1.6.1
pip        21.3.1
setuptools 39.2.0
